Pulled out actions/reducers and added an index
