+++ src/client/src/combineReducers.ts
@@ -4,12 +4,12 @@ import { connectionStatusReducer } from './connectionStatusOperations'
 import { currencyPairReducer } from './currencyPairsOperations'
 import { enviromentReducer } from './enviromentReducer'
 import { pricingServiceReducer } from './pricingOperations'
-import { analyticsReducer } from './ui/analytics/index'
-import { blotterReducer } from './ui/blotter/index'
+import { analyticsReducer } from './ui/analytics'
+import { blotterReducer } from './ui/blotter'
 import { regionsReducer } from './ui/common/regions/regionsOperations'
 import footerReducer from './ui/footer/FooterOperations'
-import { sidebarRegionReducer } from './ui/sidebar/SidebarRegionOperations'
-import { spotTileDataReducer } from './ui/spotTile/index'
+import { sidebarRegionReducer } from './ui/sidebar'
+import { spotTileDataReducer } from './ui/spotTile'
 import notionalsReducer from './ui/spotTile/notional/NotionalOperations'
 
 const rootReducer = combineReducers({
