+++ src/client/src/ui/sidebar/actions.ts
@@ -9,19 +9,3 @@ export const SidebarRegionActions = {
 }
 
 export type SidebarRegionActions = ActionUnion<typeof SidebarRegionActions>
-
-export type SidebarRegionState = boolean
-
-const initialState: SidebarRegionState = true
-
-export const sidebarRegionReducer = (
-  state: SidebarRegionState = initialState,
-  action: SidebarRegionActions
-): SidebarRegionState => {
-  switch (action.type) {
-    case ACTION_TYPES.TOGGLE_ANALYTICS:
-      return !state
-    default:
-      return state
-  }
-}
