+++ src/client/src/ui/sidebar/index.ts
@@ -1,0 +1,2 @@
+export { default } from './SidebarRegionContainer'
+export { sidebarRegionReducer } from './reducer'
