+++ src/client/src/ui/sidebar/reducer.ts
@@ -1,0 +1,17 @@
+import { ACTION_TYPES, SidebarRegionActions } from './actions'
+
+export type SidebarRegionState = boolean
+
+const initialState: SidebarRegionState = true
+
+export const sidebarRegionReducer = (
+  state: SidebarRegionState = initialState,
+  action: SidebarRegionActions
+): SidebarRegionState => {
+  switch (action.type) {
+    case ACTION_TYPES.TOGGLE_ANALYTICS:
+      return !state
+    default:
+      return state
+  }
+}
